Split code properly
