+++ typescript/web/src/components/labeling-tool/openlayers-map/draw-interaction/index.tsx
@@ -1,7 +1,7 @@
 import { MutableRefObject } from "react";
 import { useRouter } from "next/router";
 import { useLabelingStore, Tools } from "../../../../connectors/labeling-state";
-import { DrawIogInteraction } from "./iog";
+import { DrawIogInteraction } from "../iog";
 import { DrawBoundingBoxAndPolygonInteraction } from "./bounding-box-polygon";
 
 export const DrawInteraction = ({
